BIG CHANGES
Added the ability to create and remove files/directories, changed page_contents function to output as a dictionary with a word frequency count. Writes to individual text files that have the frequency count in the contents and the word as a filename. Directory name is the title of the page